Move levels and lines, fix missing atom data property
